feat: special case handling for GeoTiff w/ single Grayscale band
Adds pre-processing in case a file w/ one or two bands is detected,
where the first band is a Grayscale band. If the second band exists it
is assumed to be the Alpha band. Otherwise the pre-processing tries to
create an Alpha band. Afterwards the image is converted to an RGBA
image.

ING-2008